test: remove remote usage from chromium specs
