+++ spec/static/main.js
@@ -197,21 +197,6 @@ ipcMain.on('set-client-certificate-option', function (event, skip) {
   event.returnValue = 'done'
 })
 
-ipcMain.on('create-window-with-options-cycle', (event) => {
-  // This can't be done over remote since cycles are already
-  // nulled out at the IPC layer
-  const foo = {}
-  foo.bar = foo
-  foo.baz = {
-    hello: {
-      world: true
-    }
-  }
-  foo.baz2 = foo.baz
-  const window = new BrowserWindow({ show: false, foo: foo })
-  event.returnValue = window.id
-})
-
 ipcMain.on('prevent-next-will-attach-webview', (event) => {
   event.sender.once('will-attach-webview', event => event.preventDefault())
 })
